Refactor core OF implementation to use a domain, remove eval wrapper
