+++ packages/grafana-runtime/package.json
@@ -60,7 +60,7 @@
     "@grafana/ui": "12.4.0-pre",
     "@openfeature/core": "^1.9.0",
     "@openfeature/ofrep-web-provider": "^0.3.3",
-    "@openfeature/web-sdk": "^1.6.1",
+    "@openfeature/react-sdk": "^1.2.0",
     "@types/systemjs": "6.15.3",
     "history": "4.10.1",
     "lodash": "^4.17.23",
