+++ packages/grafana-runtime/src/internal/index.ts
@@ -28,7 +28,8 @@ export {
 
 export { UserStorage } from '../utils/userStorage';
 
-export { initOpenFeature, evaluateBooleanFlag } from './openFeature';
+export { initOpenFeature, getFeatureFlagClient } from '../internal/openFeature';
+
 export { getAppPluginMeta, getAppPluginMetas, setAppPluginMetas } from '../services/pluginMeta/apps';
 export {
   useAppPluginMeta,
