+++ packages/grafana-runtime/src/internal/openFeature/index.ts
@@ -1,5 +1,5 @@
 import { OFREPWebProvider } from '@openfeature/ofrep-web-provider';
-import { OpenFeature } from '@openfeature/web-sdk';
+import { OpenFeature } from '@openfeature/react-sdk';
 
 import { FeatureToggles } from '@grafana/data';
 
@@ -7,15 +7,16 @@ import { config } from '../../config';
 
 export type FeatureFlagName = keyof FeatureToggles;
 
-export async function initOpenFeature() {
-  /**
-   * Note: Currently we don't have a way to override OpenFeature flags for tests or localStorage.
-   * A few improvements we could make:
-   * - When running in tests (unit or e2e?), we could use InMemoryProvider instead
-   * - Use Multi-Provider to combine InMemoryProvider (for localStorage) with OFREPWebProvider
-   *   to allow for overrides https://github.com/open-feature/js-sdk-contrib/tree/main/libs/providers/multi-provider
-   */
+// The domain creates the unique instance of the OpenFeature client for Grafana core,
+// with its own evaluation context and provider.
+// Plugins should not use this client or domain, and instead create their own client
+// with a different domain to avoid conflicts.
+//
+// If changing this, you MUST also update the same constant in packages/grafana-test-utils/src/utilities/featureFlags.ts
+// to ensure tests work correctly.
+export const GRAFANA_CORE_OPEN_FEATURE_DOMAIN = 'internal-grafana-core';
 
+export async function initOpenFeature() {
   const subPath = config.appSubUrl || '';
   const baseUrl = `${subPath}/apis/features.grafana.app/v0alpha1/namespaces/${config.namespace}`;
 
@@ -25,12 +26,18 @@ export async function initOpenFeature() {
     timeoutMs: 5_000,
   });
 
-  await OpenFeature.setProviderAndWait(ofProvider, {
+  await OpenFeature.setProviderAndWait(GRAFANA_CORE_OPEN_FEATURE_DOMAIN, ofProvider, {
     targetingKey: config.namespace,
     ...config.openFeatureContext,
   });
 }
 
-export function evaluateBooleanFlag(flagName: FeatureFlagName, defaultValue: boolean): boolean {
-  return OpenFeature.getClient().getBooleanValue(flagName, defaultValue);
+/**
+ * Get the OpenFeature client for Grafana core.
+ * Prefer to instead use the React hooks for evaluating feature flags instead as they remain up to date with the latest flag values.
+ * If you must use this client directly, do not store the evaluation result for later - always call `getFeatureFlagClient().getFooValue()` just
+ * in time when you use it to ensure you get the latest value.
+ */
+export function getFeatureFlagClient() {
+  return OpenFeature.getClient(GRAFANA_CORE_OPEN_FEATURE_DOMAIN);
 }
